implementação hash encadeado multiplicativo
Implementação de função hash utilizando função de multiplicação, aplicada em uma tabela de hash encadeado.
Consiste nos métodos:
hash: calcula o indice dentro da tabela através de função de multiplicação usando proporção aurea como indice multiplicativo
inserir: insere um registro dentro da tabela. Em caso de colisão, coloca-o dentro da lista encadeada
busca: através do indice calculado no hash, busca um registro dentro da tabela.
main: permite a execução do código, tendo 3 modos de execução: 
[1] 100000 registros em uma tabela de tamanho 10000
[2] 1000000 registros em uma tabela de tamanho 100000
[3] 10000000 registros em uma tabela de tamanho 1000000

Além disso, são geradas métricas para analise: Tempo total de inserção e busca, além do total de colisões que ocorreram na inserção. Essas métricas são salvas em um CSV para analise posterior

+++ TabelaHashEncadeada.java
@@ -18,10 +18,10 @@ public class TabelaHashEncadeada {
     }
 
     // metodo inserir
-    public int inserir(Registro registro) {
+    public long inserir(Registro registro) {
         int chave = registro.obterCodigo();
         int indice = hash(chave);
-        int colisao = 0;
+        long colisao = 0;
 
         // verificar se da pra inserir no vetor
         if (vetorHash[indice] == null) {
@@ -62,7 +62,7 @@ public class TabelaHashEncadeada {
         long semente = 12345L;
         int[] arrayQuantidadeRegistro = {100_000, 1_000_000, 10_000_000};
         int[] arrayTamanhos = {10_000, 100_000, 1_000_000};
-        int colisoes = 0; // contador de colisões
+        long colisoes = 0; // contador de colisões
         int quantidadeRegistros = -1;
         int tamanhoTabelaHash = -1;
